Fix numerous errors revealed by -Wall
It's a shame perl's default doesn't include -Wall.

This corrects a major bug where I was exiting a function
without 'return', which honestly should be a fatal gcc
compile error.  grr.

Another significant bug was constructing Picture objects
with the ->{planes} attribute, which was using the ref as
a buffer instead of the SV it points to.